fix(dspark): moe_ep single-card path — fp32 grouped-GEMM (dtype mismatch)
moe_dispatch_ep degenerate branch (no EP / single rank on NPU) passed bf16
x[tok] straight to the fused NPU grouped-GEMM, which returns bf16, then
index_add_ into an fp32 accumulator -> "self (Float) and source (BFloat16)
must have the same scalar type". The EP (multi-rank) path already floats its
input (xf = x[tok].float()), and the CPU ref floats internally, so only the
single-card/no-EP NPU path was affected -- i.e. the forward-only router probe;
real EP training / conversion / vllm inference never hit this branch.

Float the input so the whole local grouped-GEMM runs fp32 (matches EP + CPU),
plus a belt out.float() before the accumulate.

+++ src/speculators/models/dsv4_dspark/backbone/moe_ep.py
@@ -129,9 +129,14 @@ def moe_dispatch_ep(x: torch.Tensor, weights: torch.Tensor, indices: torch.Tenso
     # ---- degenerate: no EP (single rank / unconfigured) == grouped over all local experts ----
     if ep is None or ep.size == 1 or not dist.is_initialized():
         n_local = ep.experts_per_rank if ep is not None else n_routed_experts
-        out = _local_grouped_ffn(x[tok], eid, w, experts, n_local)
+        # Float the input so the NPU fused grouped-GEMM runs (and returns) fp32 -- matching the
+        # EP path (line ~142 `xf = x[tok].float()`) and the CPU ref. Without this the fused NPU
+        # kernel returns bf16 and the fp32 index_add_ below raises a scalar-type mismatch (the
+        # single-card / no-EP NPU path -- e.g. the forward-only router probe -- was the only
+        # caller of this branch, so it went unnoticed).
+        out = _local_grouped_ffn(x[tok].float(), eid, w, experts, n_local)
         y = torch.zeros(T, dim, dtype=torch.float32, device=device)
-        y.index_add_(0, tok, out)
+        y.index_add_(0, tok, out.float())  # belt: no-op when already fp32
         return y
 
     # ---- route each unit to its expert's owner rank, sort by owner ----
